fix: s_nonintrasearch format

+++ mobile_insight/analyzer/lte_rrc_analyzer.py
@@ -403,7 +403,7 @@ class LteRrcAnalyzer(ProtocolAnalyzer):
                     self.profile.update("LteRrcProfile:"+str(self.__status.id)+"_"+str(self.__status.freq)+".idle.serv_config",
                                         {'priority':field_val['lte-rrc.cellReselectionPriority'],
                                          'threshserv_low':str(int(field_val['lte-rrc.threshServingLow'])*2),
-                                         's_nonintrasearch':str(int(field_val['lte-rrc.s_NonIntraSearch'])*2),
+                                         's_nonintrasearch':str(float(field_val['lte-rrc.s_NonIntraSearch'])*2),
                                          'q_hyst':field_val['lte-rrc.q_Hyst']})
 
                 self.__config[cur_pair].sib.intra_freq_config = LteRrcSibIntraFreqConfig(
